Copy PID project as starting point for MPC
Copied from https://github.com/ericlavigne/CarND-PID-Clojure
as starting point for MPC project. I expect the communication
protocol to be similar. The PID controller may also be useful
as part of the policy function, with the idea that the best
steering angle will be close to the steering angle that the
PID controller would choose.