Define frenet coordinate system based on waypoints
Created mpc.frenet namespace to handle representation and
conversion of frenet coordinates. Created track function
to create a "track" around a series of waypoints in which
s represents distance along the track and d represents
distance across track (to the right). Created sd-to-xy
function which converts (s,d) coordinates to (x,y)
coordinates in the context of a track.