Use proportional controller based on frenet d
Created a frenet (s,d) coordinate system based on the waypoints
and used d as the cross-track error for a proportional controller.
The car steers reasonably well around the first turn, but has the
unexplained behavior of driving on the left side of the road
(instead of the expected center-of-road driving) after the first
turn.