Calculate and display a value estimate for car's path
Created a value function that estimates the desirability of a
state, with the expectation that this result will be averaged or
summed over all states in a path. Value contributors include
progress along the road, distance from road center, alignment
with direction of the road, and whether the car is on the road.
(There is a very large penalty for going off-road.) For now,
the summed value for the car's plan is printed to the console.
Soon that value will be used to decide between alternative
plans.